Cleaning up exceptions

+++ python/GangaLHCb/Lib/RTHandlers/GaudiExecRTHandlers.py
@@ -335,6 +335,7 @@ def uploadLocalFile(job, namePattern, localDir, should_del=True):
     trySEs = getConfig('DIRAC')['allDiracSE']
     random.shuffle(trySEs)
     new_lfn = os.path.join(getInputFileDir(job), namePattern)
+    returnable = None
     for SE in trySEs:
         #Check that the SE is writable
         if execute('checkSEStatus("%s", "%s")' % (SE, 'Write')):
@@ -342,7 +343,9 @@ def uploadLocalFile(job, namePattern, localDir, should_del=True):
                 returnable = new_df.put(force=True, uploadSE=SE, lfn=new_lfn)[0]
                 break
             except GangaDiracError as err:
-                raise GangaException("Upload of LFN %s to SE %s failed" % (new_lfn, SE)) 
+                raise GangaException("Upload of input file as LFN %s to SE %s failed" % (new_lfn, SE)) 
+    if not returnable:
+        raise GangaException("Failed to upload input file to any SE")
     if should_del:
         os.unlink(os.path.join(localDir, namePattern))
